moved util tests
