+++ src/applications/health-care-questionnaire/shared/utils/tests/appointment.id.from.url.unit.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import { getAppointmentIdFromUrl } from '../utils';
+import { getAppointmentIdFromUrl } from '../index';
 
 describe('health care questionnaire -- utils -- get id from url in window', () => {
   it('window is undefined', () => {
